refactor(zoom): Add zoom as directive
Add zoom behavior as directive to the graph.
Capsulates the zoom behavior to be able to add it optionally to a graph.
